Anpassungen Skripte um H2 Tests nicht mehr zu reissen

+++ src/main/resources/db/migration/V3__createstadiumstable.sql
@@ -1,6 +1,5 @@
--- db/migration/V3__createstadiumstable.sql
 CREATE TABLE IF NOT EXISTS stadiums (
-                                        id SERIAL PRIMARY KEY,
+                                        id BIGINT PRIMARY KEY,
                                         name VARCHAR(255) NOT NULL,
                                         location VARCHAR(255) NOT NULL,
                                         capacity INT NOT NULL
